Updated openstack/openstack Project: openstack/heat  d6d836ff6226e0dec21c9658bf72a629b85995bf
Add the HOT fuction str_replace_vstrict

The already existing str_replace_strict function raises an error if a
param is not present in the template. str_replace_vstrict, newly added
in this patch, is identical but also raises an error if any of the
params have an empty value.

Change-Id: I5407135cc0435cfbad2d18964fe2119c999f67a3