Updated openstack/openstack Project: openstack/heat  320c820ed8b8c20b35d751a14758d329bb010648
Explicitly compare IDs to None

Many database IDs (Resource, WatchRule) are integers, so 0 is a valid ID
and evaluating it in a boolean context may give the wrong result for the
first entry in the table. Even for the ones that aren't (e.g. Stack) it's
better to explicitly compare to None.

Change-Id: Id95eeb5cd0f9cac82937fae48122ce5fa90ca48d