Updated openstack/openstack Project: openstack-infra/tripleo-ci  87722b14c7f000e0f35337bf8a82bb347806b45f
Ensure consistent path for multinode upgrade job

Some changes in t-h-t recently mean we've got to support multinode
scenario environments that reference j2 rendered files, but this
only works when the environment is contained within the --templates
tree.  So we copy the environment under test to the "new" tree
during an upgrade test, and ensure all paths reference the new
location.

Change-Id: I0288c2b470606d990281f6c37071b632a2d0d423
Closes-Bug: #1687908