Updated openstack/openstack Project: openstack/nova-specs  b4192b693b057d52e4b81990e858ffd91bb982de
update flavor info in server details spec

At the Summit Matt Riedemann pointed out that rxtx_factor was going
to get removed in the not-too-distance future, and so we probably
don't want to enshrine it in the flavor details here.

This updates the spec accordingly (and fixes a spelling nit).

APIImpact
blueprint: instance-flavor-api

Change-Id: I43680ea8ddc36ce55907d39710e6cb6cca0e6ab0