Updated openstack/openstack Project: openstack-infra/project-config  9104baa1934a651b2897943fdad255dd3fb337e3
Create new Barbican gates that have removed unused services

For the barbcan functional tests, most of the core
services such as Cinder and Nova are not needed.
Hopefully we can remove them and speed up the gates.
This change adds them as experimental checks. If they succeed, we can
promote these to be the main gates and remove the existing ones.

Change-Id: I68810330dbee4033f8198f39aba5b75cd3357399