Updated openstack/openstack Project: openstack-infra/tripleo-ci  14da18f2e7a902b252f90711cc80935d3bc34305
Guard adding bridge interface with $STABLE_RELEASE

Adding the dummy interface for the br-ctlplane bridge is only needed for
the mitaka branch where the net_config_override feature is not present.
Adding the guard and also a check if the interface has already been
previously added in case running locally.

Change-Id: I88e40470c1490745a4b69870c94f0a5f31dd89cb