Updated openstack/openstack Project: openstack-infra/shade  0ba80fa8c6e8a4c3b1b30b537d3da80fd1c393a9
Convert first ironic_client test to REST

Add the various plumbing and convert one test.

Also, ironicclient prepends "v1" to all of its calls (joy) so there is
an ugly transition related to version discovery coming with this one.
For now just prepend it in our mock base class. When we convert calls,
we'll have to prepend v1 at the call site until we're done with the
transition. Then we can do a followup patch that'll do discovery and
remove the v1 from the code. Gross.

Change-Id: Icfd70267910d19a7da2ae8e9b9e4f24d2077b87d