Updated openstack/openstack Project: openstack-infra/devstack-gate  49bde6766801c9810a1f07e0cff42b81be50b4ab
bump devstack-tools to 0.4.0

In working through converting neutron to use devstack-tools for
functional tests, we discovered that merging post-config sections had
some edge cases when the section in question happened at the end of
the file. That was fixed in 0.4.0 with tests. We should rev that as
our base file.

Change-Id: Ia1291539cf03d1094f02c5226ffc9692b6afdf02