Updated openstack/openstack Project: openstack/nova  722e48fac435e197eedc7c30806a61034ad3cc82
Make compute auto-disable itself if builds are failing

This implements an auto-disable feature in nova-compute, where we
automatically set our service record to disabled if we consecutively
fail to build a certain number of instances.

While this is a very useful thing to do in general, disabling a failing
compute becomes more important in the future where scheduler retries due
to unknown failures may become either impossible or scoped to a single
cell. Since a compute that is consistently failing will look very
attractive to the scheduler, it may become a build magnet, that in the
absence of retries, would effectively kill all builds in a cloud until
fixed.

Change-Id: I02b7cd87d399d487dd1d650540f503a70bc27749