Updated openstack/openstack Project: openstack/governance  fbed10f1953d56cdec4d7412cd6693a10169f3d3
Change the target for this goal to uWSGI not Apache mod_wsgi

Apache mod_wsgi has substantial downside in use for development, the
devstack project is pushing forward with uWSGI by default (which is
now running in the gate). This is actually *much* easier on projects
to configure, and makes for a more consistent developer experience
when working on projects with both wsgi and worker components.

Change-Id: I81e4e5f37c0e0679df9a275e3377b4652d814740