Updated openstack/openstack Project: openstack/keystone  ab2928f2cd11a35c06da939575c19872a0bbbecd
Handle NotFound when listing role assignments for deleted users

Keystone can use an external identity store for the users, and
store assignments for these users in the SQL database that it
manages. When a user has been deleted directly in the external
identity store, these assignments will persist. Therefore when
listing role assignments and asking for names to be included,
keystone will try to get information of the user and fail with
NotFound.

This catches the NotFound exception of the get_user and get_group
calls and fills the user values with and empty string.

Change-Id: Iec3e12f6cd1402e1e3f192b0ede5d608bd41ca1d
Closes-Bug: 1684820