Updated openstack/openstack Project: openstack/neutron  d6ba40eb4a03136ee9c915dcd347934f915f8df0
Fix tempest test failing with segments extension

This patchset fixes tempest test case test_filtering_shared_subnets that
fails when executed with the segments extension enabled, due to the fact
that the 'segment_id' attribute is returned in subnets listed with an
admin client but not in subnets listed with a non admin client.

Patchset https://review.openstack.org/#/c/459439/ proposes to enable the
segments extension in the gate.

Change-Id: I6b9bd6fe944c9485c21d25b37c943f3bd53fd7c9
Closes-Bug: #1690598