Updated openstack/openstack Project: openstack-dev/pbr  c02406601a7d7c453c77318ff5b5af6d760e80ae
builddoc: uses the new Sphinx 1.6 code for multiple builder

As noted by stephenfin, Sphinx 1.6 provides its own code to build doc with
multiple builders. The one provided by pbr so far for Sphinx < 1.6 is not even
compatible with 1.6. This patch fixes that by running the native Sphinx code
for Sphinx > 1.6 and falling back to the old code for older Sphinx versions.

Closes-Bug: #1691129
Change-Id: I5224235b1056a248b246c54e2d99eea94d53c4eb