Updated openstack/openstack Project: openstack/neutron  a6d5c396e8b42b82ab4a7079cb762f20c668249d
Add check for Bandwidth Limit Rules

Currently it is possible to create 2 rules in the same policy
where the Max Bandwidth ( from the Bandwidth Limit Rule ) can
be less than Minimum Bandwidth defined in ( Minimum Bandwidth
Rule) , which can be pretty confusing.
This patch raises an exception if such an issue occurs.

Change-Id: Ib748947bcd85253aa22e370a56870afbfbafa19b
Closes-Bug: #1667138