Updated openstack/openstack Project: openstack/tripleo-heat-templates  ea1008f09ffdf7975640491f08eed1bf1e2194fa
Use neutron-server image for neutron services

For TLS everywhere, neutron-server needs httpd in the image, since
it'll use a separate container that runs a TLS proxy to terminate
the connection. This requires the image where the configuration is
ran to have httpd installed, since there are several directories
and the user/group that's needed.

So, we then switch the image to be used to be neutron-server instead
of the openvswitch-agent image.

Change-Id: Ie16de3004925b7624f106d6c015ec04ef6031a06
Depends-On: I82f10ac0e7e692e6ba4a06dc10da9eaf79c60e7e