Updated openstack/openstack Project: openstack/ironic  9fc3a376cadb953b8df819d906805c39071937be
Check if sort key is allowed in API version

This checks the sort key, to make sure the specified field
is allowed/available in the specified API microversion.
If it is not allowed, a 406 HTTP status is returned.

This affects requests to get lists of nodes, port groups,
and ports.

Change-Id: Id5fb44b8b7fe989514dbae4b60cef4a34d47e52b
Closes-Bug: #1659419