Updated openstack/openstack Project: openstack/nova  77b11a9e31e265c338857326418bfe8cd135c1f8
Totally freeze the extension_info API

The extension will be removed from the Nova API totally. The extension_info
API is just for backward compatibility. And there is no way to disable/enable
the extension, also the discovery policy rules are removed, so it is time
to just hard-code most of the logic at here.

Partial-implement-blueprint api-no-more-extensions-pike

Change-Id: I56d859beb675199b209587bfe0a23cd9b75233ad