Updated openstack/openstack Project: openstack/heat  855d67b51ab44bc58ddcbf390469d54b840738ea
Yield to other greenthreads during validation

Validating a large stack locks up a whole worker until it is complete,
because it doesn't co-operate in greenlet's co-operative multithreading.
With this change, we now yield the flow of control between each resource,
so that validation can't monopolise the process.

Validation is high priority because it happens synchronously, but it is
probably worth the trade-off to avoid blocking anything else happening on
the same worker.

Change-Id: I8514640135a915791a0905eea40ef70876ae1cb1