Updated openstack/openstack Project: openstack-infra/project-config  541ae2f178da11897d8b265687b910e5f8fd3712
Install test packages in tempest-plugin-sanity job

Currently the requirements of in-tree tempest plugins are
kept in test-requirements.txt of each project. On doing
pip install <project>, will just install all the packages
required for that project and on listing tempest plugin
tests in a tempest workspace will give dependency error
and tempest-plugin-sanity job will fails. In order to avoid
the above failure we need to do the same.

Change-Id: I84a6c4feb11e28e98edf86e59ff0813eba856426