Updated openstack/openstack Project: openstack/diskimage-builder  ca043483939a47431eb116ca97a17be5290299fe
Remove _config_error thrower

"log and throw" is arguably an anti-pattern; the error message either
bubles-up into the exception, or the handler figures it out.  We have
an example where this logs, and then the handler in blockdevice.py
catches it and logs it again.

Less layers is better; just raise the exception, and use log.exception
to get tracebacks where handled.

Change-Id: I8efd94fbe52a3911253753f447afdb7565849185