Updated openstack/openstack Project: openstack-infra/project-config  4a36c1af81d492393bf07bf155e62e63c822eb34
Check for test-requirements.txt file then install

Some of the projects having tempest-plugin like aodh does not
have test-requirements.txt file which leds to the failure of
tempest-plugin-sanity job while installing test packages from
it. In order to avoid the failure we need to check for that
file then install.

Change-Id: I09f2ab6d2187f159f7ddb0068111ac276ca7bbcf