Updated openstack/openstack Project: openstack-infra/irc-meetings  ee13502e45eb705c75c670189d4c1ea67316bba4
Remove the defunct nova live migration meeting

The last meeting was March 28 2017 and Timofey Durakov
is no longer working on OpenStack, so there is no one
chairing this meeting anymore, and no one stepping into
it. It served a useful purpose for a lot of good work over
the last few releases (since Liberty really) but it's time
to remove this meeting so others can use this time slot
if needed.

Change-Id: Ic8747f9ccc306437be2c601b5cfaa854f109df14