Updated openstack/openstack Project: openstack-infra/system-config  6e9361a7c60155326a089697a8eed984793568b1
Add citycloud private networks, subnets, and routers

This should give us connectivity to the outside world with NAT'd
internal IP addressing.

Note that we can't add the router to the template because the external
network name will be different across clouds and we have to pass in the
subnet lists which may vary as well.

Change-Id: Iea225c71d0d8e644cbaf709554d02d130ad21c18