Updated openstack/openstack Project: openstack/tripleo-heat-templates  f231f6a3d38b06630aaba11cec3bd32b4d1ca82c
Consolidate writing json configs with HostPrepTasks

We already have an ansible deployment that applies the per-service
host_prep_tasks, so we can simplify the dependencies here by just
doing the docker-steps host preparation at the same time.

The motivation behind this is to both simplify the depends_on web we
have here, reduce the number of discrete deployments, and also to
potentially make running ansible directly e.g for debugging easier.

In a future patch we'll convert the configuration steps to work in
a similar way, such that they can be more easily reapplied e.g for
rolling minor updates, possibly outside of heat.

Change-Id: I9a201fc5a9e82c7fba4c2de36eb5332e21a81d37