Updated openstack/openstack Project: openstack/requirements  f562496f7d7b5f2978d8f35cc738122666e45f1e
Fixed wrong regular expression for vcs requirements

The regular expression to parse requirement line is also used for
constrains files. However, it will not match those lines which
point to VCS, such as git+http or git+https which pypi accept.

This commit fixes the issue.

Despite that we only need to change the RE to
^(?P<url>\s*(?:-e\s)?\s*(?:(?:git\+)?(?:https|http)|file)://[^#]*)#
egg=(?P<name>[-\.\w]+)
to fix the issue for git+http and git+https, pypi accepts more VCS
URI schemes other than git. This patch also looses the match for
these cases to be accepted.

Change-Id: I202078f5ca4ba6da4ab0ed52d97d41997674c1b4