Updated openstack/openstack Project: openstack/neutron  a91220599b23b7cdc527b0174c6896748c1ca79e
Changing create_server to be "non class method"

The change is done in order to remove VM in the test level
and not class level.

As it makes sense to have several tests in the same class and
having in mind that the resources could be limited, it makes
sense to remove unneeded resources in the end of every test.

The create_and_associate_floatingip function was changed to become
instance method as well and the creation of servers in floatingIP
scenario test was taken out from resource_setup.

Change-Id: Ibd8abc2b44b2e85de3e4cd2928799768ad1c0af5