Updated openstack/openstack Project: openstack/heat  46e42c5184166a4943cf9ed83cafca06c8de2623
Add os_collect_config attribute to server resources

Add a new attribute, os_collect_config to the server resources. The
attribute's value is a map representing the configuration for the
os-collect-config agent on the actual servers.

Adding this data as an attribute will allow it to be represented in a
stack output and also queried directly in the templates.

implements blueprint split-stack-default
Change-Id: I8acfd67cd8138d587cc362184c84a08134bf3157