Updated openstack/openstack Project: openstack-infra/project-config  175edd409b2d2b034f8d69e7e51edfff13037215
Use all available citycloud quota in happy regions

This increases max servers to 50 in three of the four citycloud regions.
Everything was happy at lower bound in Kna1, Lon1, and Sto2 so increase
max-servers to our total quota to take advantage of this new capacity.

We will increase La1 when we sort out the flavor issue there.

Change-Id: Ic7dd5873a862e2b197c8e25d4cfb06f2957f0f9a