Updated openstack/openstack Project: openstack/nova  931c3f48188e57e71aa6518d5253e1a5bd9a27c0
Correct the wording about filter options

The current docs on the Filter Scheduler imply that filters can be
specified per request. This is not the case, so this change corrects
that page to remove this incorrect language, and make clear that only
the filters specified in the config file will be used.

Closes-Bug: #1689339

Change-Id: I3459a894481cab7d3961ef875015e2b8011e9a58