Updated openstack/openstack Project: openstack/cinder  91364f97f084ecab0a77be7ee90ee55aba058341
qemu_img_info: Don't autodetect source format

Use the format information provided by Glance.

Autodetection of formats is almost never what we want
to do: it introduces possibilities for dangerous
behavior and complexity, and we can avoid it here
because Glance gives us enough information to not need
it.

More directly for the encrypted image->volume case,
this prevents qemu-img from treating an encrypted image
as a "luks" container that it should decrypt, instead
treating it as raw data.

Change-Id: I4071927c491870626fe174b75ecaf8ef6da39cf5