Updated openstack/openstack Project: openstack-infra/shade  e3d43e7396d95d6cfaa4ca0113108cbef22225b6
Find private ip addr based on fip attachment

In the case of multiple private ip addresses we previous returned the
first private ip address found. That can cause problems if that IP
address isn't valid. Since the IPs are private the only real validity
check we can use is if the private IP address has a floating IP
attaached. Update shade to return the private IP that is attached to a
floating IP when there are multiple private IPs.

Change-Id: I86807c5149dfad72948ff6832a6697b594d1f195