Updated openstack/openstack Project: openstack/nova  63f2b761592003b9ade2c6902556e186253e1852
XenAPI: remove hardcode dom0 plugin version in unit test

When XenAPI driver are calling dom0 plugin via os-xenapi, it will
check whether the dom0 plugin's version is match which is fine.
But our FakeXenAPISession is using the hardcode dom0 plugin version
for unit test which make the UT fail each time we bump dom0 plugin
version in os-xenapi repo. This patch is to remove the hardcode
and use the exact version from os-xenapi

Change-Id: I52939fbd42b1b1f7f6c216f3715158b779c597ab