Updated openstack/openstack Project: openstack/tripleo-heat-templates  9a0d017cc74cc8713d3b1d9ae8172e2d7d0611ce
Use disabled suffix for disabled service names

This changes both the service names and the file names for disabled
services, adding the 'disabled' suffix to them.

This comes with the reasoning that, if a service requires a disabled
service, and checks for the name in the "service_names" hiera entry, it
will appear as if the service was enabled, when it's actually not. So
changing the name and using this convention prevents that issue.

Change-Id: I308d6680a4d9b526f22ba0d7d20e5db638aadb9a