Updated openstack/openstack Project: openstack/tripleo-heat-templates  ecc231438850a057e3dca8bc9673c9e22d54213d
docker/TLS-everywhere: Add metadata_settings output to templates

This is needed since it's what writes the service metadata to the nova
server in order to create the kerberos principals. It worked in a base
controller since the keystone template does have this. But if we would
deploy these services on a separate role, it would break. So this output
is needed.

bp tls-via-certmonger-containers

Change-Id: I3ee8c65d356dcd092a3fbf79041e5c69ef23b721